requirements.txt
# Core dependencies
requests>=2.28.1
beautifulsoup4>=4.11.1
scapy>=2.4.5
python-nmap>=0.7.1
cryptography>=38.0.1
psutil>=5.9.3
pyopenssl>=22.1.0

# Wireless attack dependencies
pyric>=0.1.6
wireless>=0.3.3

# NOTE: For Bluetooth functionality on non-Windows systems, install pybluez manually:
# pip install git+https://github.com/pybluez/pybluez.git#egg=pybluez

# Web security testing dependencies
lxml>=4.9.1
numpy
sslyze>=5.1.0
dnspython>=2.2.1  # For DNS operations
defusedxml>=0.7.1  # For safe XML processing

# Exploitation dependencies
impacket>=0.10.0
requests-ntlm>=1.1.0
ldap3>=2.9.1
pyyaml>=6.0
pwntools>=4.9.0

# Optional dependencies for enhanced functionality
pillow>=9.3.0  # For image processing and banner generation
pyserial>=3.5  # For hardware interfaces
flask>=2.2.2  # For API interfaces
pyinstaller>=5.6.2  # For creating standalone executables
pywin32>=305; platform_system == "Windows"  # For Windows-specific functions

# Documentation dependencies
mkdocs>=1.5.0  # Documentation generator
mkdocs-material>=9.4.0  # Material theme
mkdocstrings>=0.23.0  # API documentation
mkdocstrings-python>=1.7.0  # Python support
mkdocs-git-revision-date-localized-plugin>=1.2.0  # Git integration
mkdocs-git-committers-plugin  # Git contributors

# AI module dependencies
numpy>=1.23.0  # Required for AI vulnerability analyzer
scikit-learn>=1.1.0  # For machine learning functionality
tensorflow>=2.10.0; platform_system != "Windows" or platform_machine != "ARM64"  # Optional AI backend

pytest
pytest-cov